refactor(inbox): extract shared copyInboxReportLink helper
Dedupes the identical copy-deeplink handler across the report, run, and
pull request detail views into a single utility.
